% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_file.R
\name{execute_file}
\alias{execute_file}
\title{Performs data acquisition from local file}
\usage{
execute_file(file, timestamp = FALSE)
}
\arguments{
\item{file}{\link{character} (\strong{required}): Path or file to be imported.}

\item{timestamp}{\link{logical} (\strong{optional}): Is a timestamp available in the data?}
}
\description{
Performs data acquisition from local file
}
\examples{
##=====================================
## Example 1: read data from local file
##=====================================

file <- system.file("extdata", "short_y_impulse.csv", package="HypeRIMU")

data <- execute_file(file, timestamp = TRUE)

}
