% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allSensors.R
\name{get_allSensors}
\alias{get_allSensors}
\title{Extract sepcific sensor from data}
\usage{
get_allSensors(data)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}): \link{data.frame} from \link{execute_file} or \link{execute_TCP}.

Note: Case sensitive}
}
\description{
Extract sepcific sensor from data
}
\examples{
##=====================================
## Example 1: extract sepcific sensor
##=====================================

file <- system.file("extdata", "short_y_impulse.csv", package="HypeRIMU")

data <- execute_file(file, timestamp = TRUE)

MPL_Linear_Acceleration <- get_allSensors(data)


}
