% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specificSensor.R
\name{get_specificSensor}
\alias{get_specificSensor}
\title{Extract sepcific sensor from data}
\usage{
get_specificSensor(data, sensorName)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}): \link{data.frame} from \link{execute_file} or \link{execute_TCP}.}

\item{sensorName}{\link{character} (\strong{required}): Which sensor should be extracted?
Allowed sensor names: MPL_Gyroscope, MPL_Accelerometer, MPL_Magnetic_Field, MPL_Orientation,
MPL_Rotation_Vector, MPL_Game_Rotation, MPL_Linear_Acceleration, MPL_Gravity, MPL_Significant_Motion,
MPL_Step_Detector, MPL_Step_Coutner, MPL_Geomagnetic_Rotation, CM36686_Proximity_Sensor, CM36686_Light_Sensor,
Screen_Orientation_Sensor

Note: Case sensitive}
}
\description{
Extract sepcific sensor from data
}
\examples{
##=====================================
## Example 1: extract sepcific sensor
##=====================================

file <- system.file("extdata", "short_y_impulse.csv", package="HypeRIMU")

data <- execute_file(file, timestamp = TRUE)

MPL_Linear_Acceleration <- get_specificSensor(data, "MPL_Linear_Acceleration")


}
