% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypeRIMU.R
\docType{package}
\name{HyperIMU-package}
\alias{HyperIMU-package}
\title{HypeRIMU}
\description{
A collection of function to illustrate the use of smartphone sensors
}
\details{
\tabular{ll}{ Package: \tab HypeRIMU\cr Type: \tab Package\cr Version:
\tab 0.0.1\cr Date: \tab 2017-XX-XX \cr License: \tab GPL-3\cr }
}
\author{
\bold{Authors}
\tabular{ll}{Johannes Friedrich \tab University of Bayreuth, Germany }

\bold{Support contact}

\email{johannes.friedrich@uni-bayreuth.de}\cr

\bold{Project source code repository}\cr
\url{https://github.com/JohannesFriedrich/HyperIMU}\cr

\bold{Related package projects}\cr
\url{http://www.r-luminescence.de}\cr

 \bold{Package maintainer}

Johannes Friedrich, University of Bayreuth, Germany
\cr \email{johannes.friedrich@uni-bayreuth.de}
}
\keyword{package}
