% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_file.R
\name{execute_file}
\alias{execute_file}
\title{Performs data acquisition from local file}
\usage{
execute_file(filepath, timestamp = FALSE)
}
\arguments{
\item{filepath}{\link{character} (\strong{required}): path and file to be imported.}

\item{timestamp}{\link{logical}: Is a timestamp available in the data?}
}
\description{
Performs data acquisition from local file
}
\examples{
##=====================================
## Example 1: read data from local file
##=====================================

file <- system.file("extdata", "HIMU-2017-07-25_11-26-18.csv", package="HypeRIMU")

data <- execute_file(file, timestamp = TRUE)

}
