% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_JSON.R
\name{execute_JSON}
\alias{execute_JSON}
\title{Performs data acquisition from data received from JSON}
\usage{
execute_JSON(port, timestamp = FALSE, timeout = 10)
}
\arguments{
\item{port}{\link{numeric} (\strong{required}): Number of the port}

\item{timestamp}{\link{logical} (\strong{with default}): Is a timestamp available in the data?}

\item{timeout}{\link{integer} (\strong{optional}): the timeout (in seconds) to be used for this connection.}
}
\description{
Performs data acquisition from data received from JSON
}
\examples{
\dontrun{
##==========================================
## Example 1: read data from JSON connecntion
##==========================================

data <- execute_JSON(port = 5555, timestamp = T)
## start HyperIMU app

}
}
