% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_TCP.R
\name{execute_TCP}
\alias{execute_TCP}
\title{Performs data acquisition from local file}
\usage{
execute_TCP(port, timestamp = FALSE, sensor_names = NULL)
}
\arguments{
\item{port}{\link{numeric} (\strong{required}): Number of the port}

\item{timestamp}{\link{logical}: Is a timestamp available in the data?}

\item{sensor_names}{\link{character} (\strong{optional}): Name of the exported sensors}
}
\description{
Performs data acquisition from local file
}
\examples{
\dontrun{

data <- execute_TCP(port = 5555, timestamp = T)
## start HyperIMU app

}
}
