% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypeRIMU-package.R
\docType{package}
\name{HypeRIMU-package}
\alias{HypeRIMU-package}
\title{HypeRIMU}
\description{
A collection of functions to illustrate the usage of smartphone sensors with the app HyperIMU
}
\details{
\tabular{ll}{Package: \tab HypeRIMU\cr Type: \tab Package\cr Version:
\tab 0.0.1 \cr Date: \tab 2018-08-26 \cr License: \tab GPL-3\cr }
}
\keyword{package}
